+++ TeamMeetingLogs/Release Notes.md
@@ -3,7 +3,12 @@
 
 - Created database
 - Added Login & Registration pages
-- 
+
+
+# April 7th
+
+- Added likes and comments
+- Added a page to view recipes